requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2020.6.20=hecda079_0
cairo=1.14.12=h8948797_3
certifi=2020.6.20=py37hc8dfbb8_0
chardet=3.0.4=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py37_0
dbus=1.13.16=hb2f20db_0
expat=2.2.9=he6710b0_2
ffmpeg=4.1.3=h167e202_0
fontconfig=2.13.0=h9420a91_0
freetype=2.10.2=h5ab3b9f_0
future=0.18.2=pypi_0
giflib=5.1.4=h14c3975_1
glib=2.65.0=h3eb4bd4_0
gmp=6.1.2=h6c8ec71_1
gnutls=3.6.5=h71b1129_1002
graphite2=1.3.14=h23475e2_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
harfbuzz=2.4.0=hca77d97_1
hdf5=1.10.5=nompi_h3c11f04_1104
icu=58.2=he6710b0_3
idna=2.10=pypi_0
intel-openmp=2020.1=217
jasper=1.900.1=hd497a04_4
jpeg=9d=h516909a_0
kaggle=1.5.6=pypi_0
kiwisolver=1.2.0=py37hfd86e86_0
lame=3.100=h7b6447c_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libblas=3.8.0=16_mkl
libcblas=3.8.0=16_mkl
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h63c8f33_5
liblapack=3.8.0=16_mkl
liblapacke=3.8.0=16_mkl
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libwebp=1.0.1=h8e7db2f_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=he19cac6_1
lz4-c=1.9.2=he6710b0_1
matplotlib=3.2.2=0
matplotlib-base=3.2.2=py37hef1b27d_0
mkl=2020.1=217
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.1.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
ncurses=6.2=he6710b0_1
nettle=3.4.1=hbb512f6_0
ninja=1.10.0=py37hfd86e86_0
numpy=1.19.1=py37hbc911f0_0
numpy-base=1.19.1=py37hfa32c7d_0
olefile=0.46=py37_0
opencv=4.1.0=py37h3aa1047_6
openh264=1.8.0=hdbcaa40_1000
openssl=1.1.1g=h516909a_1
pcre=8.44=he6710b0_0
pillow=7.2.0=py37hb39fc2d_0
pip=20.1.1=py37_1
pixman=0.40.0=h7b6447c_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h05f1152_2
python=3.7.7=hcff3b4d_5
python-dateutil=2.8.1=py_0
python-graphviz=0.14.1=pypi_0
python-slugify=4.0.1=pypi_0
python_abi=3.7=1_cp37m
pytorch=1.5.1=py3.7_cuda10.2.89_cudnn7.6.5_0
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
requests=2.24.0=pypi_0
scipy=1.5.2=py37hb14ef9d_0
setuptools=49.2.0=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.15.0=py_0
sqlite=3.32.3=h62c20be_0
text-unidecode=1.3=pypi_0
tk=8.6.10=hbc83047_0
torchvision=0.6.1=py37_cu102
torchviz=0.0.1=pypi_0
tornado=6.0.4=py37h7b6447c_1
tqdm=4.48.0=pypi_0
urllib3=1.24.3=pypi_0
wheel=0.34.2=py37_0
x264=1!152.20180806=h14c3975_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
